refactor: merge util files
