+++ src/components/month/calendar-month-view.component.ts
@@ -30,7 +30,7 @@ import differenceInSeconds from 'date-fns/difference_in_seconds';
 import addSeconds from 'date-fns/add_seconds';
 import { CalendarEventTimesChangedEvent } from '../../interfaces/calendar-event-times-changed-event.interface';
 import { CalendarUtils } from '../../providers/calendar-utils.provider';
-import { validateEvents } from '../common/util';
+import { validateEvents } from '../../providers/util';
 
 /**
  * Shows all events on a given month. Example usage:
