+++ src/components/week/calendar-week-view.component.ts
@@ -25,7 +25,7 @@ import { CalendarDragHelper } from '../../providers/calendar-drag-helper.provide
 import { CalendarResizeHelper } from '../../providers/calendar-resize-helper.provider';
 import { CalendarEventTimesChangedEvent } from '../../interfaces/calendar-event-times-changed-event.interface';
 import { CalendarUtils } from '../../providers/calendar-utils.provider';
-import { validateEvents } from '../common/util';
+import { validateEvents } from '../../providers/util';
 
 export interface WeekViewEventResize {
   originalOffset: number;
