+++ src/providers/util.ts
@@ -1,3 +1,13 @@
+import {
+  CalendarEvent,
+  validateEvents as validateEventsWithoutLog
+} from 'calendar-utils';
+
+export const validateEvents = (events: CalendarEvent[]) => {
+  const warn = (...args) => console.warn('angular-calendar', ...args);
+  return validateEventsWithoutLog(events, warn);
+};
+
 export function isInside(outer: ClientRect, inner: ClientRect): boolean {
   return (
     outer.left <= inner.left &&
